pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-sky"
description = "基于Nonebot2的sky光遇攻略插件"
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
authors = [
    { name="Kaguya233qwq", email="marisa_qwq@qq.com" },
]
keywords = ["nonebot2", "nonebot", "sky光遇", "光遇"]
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Natural Language :: Chinese (Simplified)",
]
dependencies = [
    "aiofiles>=24.1.0",
    "bs4>=0.0.2",
    "click",
    "httpx>=0.28.1",
    "lxml>=6.0.0",
    "nb-cli>=1.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot2>=2.4.2",
    "setuptools>=80.9.0",
]

[project.urls]
Homepage = "https://github.com/Kaguya233qwq/nonebot_plugin_sky"
Repository = "https://github.com/Kaguya233qwq/nonebot_plugin_sky"
Issues = "https://github.com/Kaguya233qwq/nonebot_plugin_sky/issues"

# 包含数据文件 
[tool.setuptools]
# 在 MANIFEST.in 文件中指定文件
include-package-data = true

[tool.setuptools.package-data]
"nonebot_plugin_sky" = ["src/resources/**/*"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["nonebot_plugin_sky*", "sky_cli*"]

[tool.setuptools_scm]

# 附加脚手架组件
[project.scripts]
sky = "sky_cli.cli:sky"
